move all `service*` tests into a single file
