--- packages/patrol/example/integration_test/service_airplane_mode_test.dart
@@ -1,12 +1,0 @@
-import 'common.dart';
-
-void main() {
-  patrol('disables and enables airplane mode twice', ($) async {
-    await createApp($);
-
-    await $.native.disableAirplaneMode();
-    await $.native.enableAirplaneMode();
-    await $.native.disableAirplaneMode();
-    await $.native.enableAirplaneMode();
-  });
-}
