--- packages/patrol/example/integration_test/service_bluetooth_test.dart
@@ -1,12 +1,0 @@
-import 'common.dart';
-
-void main() {
-  patrol('disables and enables bluetooth twice', ($) async {
-    await createApp($);
-
-    await $.native.disableBluetooth();
-    await $.native.enableBluetooth();
-    await $.native.disableBluetooth();
-    await $.native.enableBluetooth();
-  });
-}
