--- packages/patrol/example/integration_test/service_cellular_test.dart
@@ -1,12 +1,0 @@
-import 'common.dart';
-
-void main() {
-  patrol('disables and enables cellular twice', ($) async {
-    await createApp($);
-
-    await $.native.disableCellular();
-    await $.native.enableCellular();
-    await $.native.disableCellular();
-    await $.native.enableCellular();
-  });
-}
