--- packages/patrol/example/integration_test/service_dark_mode_test.dart
@@ -1,12 +1,0 @@
-import 'common.dart';
-
-void main() {
-  patrol('disables and enables dark mode twice', ($) async {
-    await createApp($);
-
-    await $.native.disableDarkMode();
-    await $.native.enableDarkMode();
-    await $.native.disableDarkMode();
-    await $.native.enableDarkMode();
-  });
-}
