--- packages/patrol/example/integration_test/service_wifi_test.dart
@@ -1,12 +1,0 @@
-import 'common.dart';
-
-void main() {
-  patrol('disables and enables wifi twice', ($) async {
-    await createApp($);
-
-    await $.native.disableWifi();
-    await $.native.enableWifi();
-    await $.native.disableWifi();
-    await $.native.enableWifi();
-  });
-}
